# app/Http/Controllers/OpportunityController.php
<?php

namespace App\Http\Controllers;

use App\Services\OpportunityService;

class OpportunityController extends Controller
{
    /**
     * @OA\Get (
     *     path="/api/opportunity",
     *     tags={"Unauthenticated"},
     *     @OA\Parameter(
     *         name="search",
     *         in="query",
     *         @OA\Schema(
     *             type="string"
     *         )
     *     ),
     *     @OA\Parameter(
     *         name="city_id",
     *         in="query",
     *         @OA\Schema(
     *             type="integer"
     *         )
     *     ),
     *     @OA\Parameter(
     *         name="state_id",
     *         in="query",
     *         @OA\Schema(
     *             type="integer"
     *         )
     *     ),
     *     @OA\Parameter(
     *         name="opportunity_type_id",
     *         in="query",
     *         @OA\Schema(
     *             type="integer"
     *         )
     *     ),
     *     @OA\Parameter(
     *         name="occupation_area_id",
     *         in="query",
     *         @OA\Schema(
     *             type="integer"
     *         )
     *     ),
     *      @OA\Response(
     *          response=200,
     *          description="success",
     *          @OA\JsonContent(
     *             type="array",
     *             @OA\Items(
     *                 type="object",
     *                 @OA\Property(
     *                     property="company",
     *                     type="string"
     *                 ),
     *                 @OA\Property(
     *                     property="cnpj",
     *                     type="string"
     *                 ),
     *                 @OA\Property(
     *                     property="state",
     *                     type="object",
     *                     @OA\Property(
     *                         property="id",
     *                         type="integer"
     *                     ),
     *                     @OA\Property(
     *                         property="title",
     *                         type="string"
     *                     ),
     *                     @OA\Property(
     *                         property="letter",
     *                         type="string"
     *                     )
     *                 ),
     *                 @OA\Property(
     *                     property="city",
     *                     type="object",
     *                     @OA\Property(
     *                         property="id",
     *                         type="integer"
     *                     ),
     *                     @OA\Property(
     *                         property="title",
     *                         type="string"
     *                     ),
     *                     @OA\Property(
     *                         property="letter",
     *                         type="string"
     *                     )
     *                 ),
     *                 @OA\Property(
     *                     property="title",
     *                     type="string"
     *                 ),
     *                 @OA\Property(
     *                     property="description",
     *                     type="string"
     *                 ),
     *                 @OA\Property(
     *                     property="full_description",
     *                     type="string"
     *                 ),
     *                 @OA\Property(
     *                     property="date",
     *                     type="string",
     *                     format="date"
     *                 ),
     *                 @OA\Property(
     *                     property="salary",
     *                     type="string"
     *                 ),
     *                 @OA\Property(
     *                     property="published_by",
     *                     type="object",
     *                     @OA\Property(
     *                         property="id",
     *                         type="integer"
     *                     ),
     *                     @OA\Property(
     *                         property="name",
     *                         type="string"
     *                     ),
     *                     @OA\Property(
     *                         property="image",
     *                         type="string"
     *                     ),
     *                     @OA\Property(
     *                         property="phone",
     *                         type="string"
     *                     ),
     *                     @OA\Property(
     *                         property="address",
     *                         type="string"
     *                     ),
     *                     @OA\Property(
     *                         property="complement",
     *                         type="string"
     *                     ),
     *                     @OA\Property(
     *                         property="state",
     *                         type="object",
     *                         @OA\Property(
     *                             property="id",
     *                             type="integer"
     *                         ),
     *                         @OA\Property(
     *                             property="title",
     *                             type="string"
     *                         ),
     *                         @OA\Property(
     *                             property="letter",
     *                             type="string"
     *                         )
     *                     ),
     *                     @OA\Property(
     *                         property="city",
     *                         type="object",
     *                         @OA\Property(
     *                             property="id",
     *                             type="integer"
     *                         ),
     *                         @OA\Property(
     *                             property="title",
     *                             type="string"
     *                         ),
     *                         @OA\Property(
     *                             property="letter",
     *                             type="string"
     *                         )
     *                     ),
     *                     @OA\Property(
     *                         property="description",
     *                         type="string"
     *                     )
     *                 ),
     *                  @OA\Property(
     *                      property="candidates",
     *                      type="array",
     *                      @OA\Items(
     *                          @OA\Property(
     *                              property="id",
     *                              type="integer"
     *                          ),
     *                          @OA\Property(
     *                              property="name",
     *                              type="string"
     *                          ),
     *                          @OA\Property(
     *                              property="email",
     *                              type="string"
     *                          ),
     *                          @OA\Property(
     *                              property="phone",
     *                              type="string"
     *                          ),
     *                          @OA\Property(
     *                              property="image",
     *                              type="string"
     *                          ),
     *                          @OA\Property(
     *                              property="created_at",
     *                              type="string",
     *                              format="date-time"
     *                          ),
     *                          @OA\Property(
     *                              property="vewed_at",
     *                              type="string",
     *                              format="date"
     *                          )
     *                      )
     *                  ),
     *                  @OA\Property(
     *                      property="type",
     *                      type="object",
     *                      @OA\Property(
     *                          property="id",
     *                          type="integer"
     *                      ),
     *                      @OA\Property(
     *                          property="title",
     *                          type="string"
     *                      )
     *                  ),
     *                  @OA\Property(
     *                      property="occupation_area",
     *                      type="object",
     *                      @OA\Property(
     *                          property="id",
     *                          type="integer"
     *                      ),
     *                      @OA\Property(
     *                          property="title",
     *                          type="string"
     *                      )
     *                  )
     *             )
     *         )
     *      )
     * )
     */
    public function index(OpportunityService $opportunityService)
    {
        return $opportunityService->list();
    }


    /**
     * @OA\Get (
     *     path="/api/opportunity/{opportunity_id}",
     *     tags={"Unauthenticated"},
     *     security={{"bearerAuth": {}}},
     *     @OA\Parameter(
     *         in="path",
     *         name="opportunity_id",
     *         required=true,
     *         @OA\Schema(type="integer")
     *     ),
     *     @OA\Response(
     *         response=200,
     *         description="success",
     *          @OA\JsonContent(
     *             type="object",
     *             @OA\Property(
     *                  property="id",
     *                  type="integer"
     *              ),
     *              type="object",
     *              @OA\Property(
     *                  property="company",
     *                  type="string"
     *              ),
     *              @OA\Property(
     *                  property="cnpj",
     *                  type="string"
     *              ),
     *              @OA\Property(
     *                  property="state",
     *                  type="object",
     *                  @OA\Property(
     *                      property="id",
     *                      type="integer"
     *                  ),
     *                  @OA\Property(
     *                      property="title",
     *                      type="string"
     *                  ),
     *                  @OA\Property(
     *                      property="letter",
     *                      type="string"
     *                  )
     *              ),
     *              @OA\Property(
     *                  property="city",
     *                  type="object",
     *                  @OA\Property(
     *                      property="id",
     *                      type="integer"
     *                  ),
     *                  @OA\Property(
     *                      property="title",
     *                      type="string"
     *                  ),
     *                  @OA\Property(
     *                      property="letter",
     *                      type="string"
     *                  )
     *              ),
     *              @OA\Property(
     *                  property="title",
     *                  type="string"
     *              ),
     *              @OA\Property(
     *                  property="description",
     *                  type="string"
     *              ),
     *              @OA\Property(
     *                  property="full_description",
     *                  type="string"
     *              ),
     *              @OA\Property(
     *                  property="status",
     *                  type="boolean"
     *              ),
     *              @OA\Property(
     *                  property="situation",
     *                  type="string"
     *              ),
     *              @OA\Property(
     *                  property="date",
     *                  type="string",
     *                  format="date"
     *              ),
     *              @OA\Property(
     *                  property="salary",
     *                  type="string"
     *              ),
     *              @OA\Property(
     *                  property="published_by",
     *                  type="object",
     *                  @OA\Property(
     *                      property="id",
     *                      type="integer"
     *                  ),
     *                  @OA\Property(
     *                      property="name",
     *                      type="string"
     *                  ),
     *                  @OA\Property(
     *                      property="image",
     *                      type="string"
     *                  ),
     *                  @OA\Property(
     *                      property="phone",
     *                      type="string"
     *                  ),
     *                  @OA\Property(
     *                      property="address",
     *                      type="string"
     *                  ),
     *                  @OA\Property(
     *                      property="complement",
     *                      type="string"
     *                  ),
     *                  @OA\Property(
     *                      property="state",
     *                      type="object",
     *                      @OA\Property(
     *                          property="id",
     *                          type="integer"
     *                      ),
     *                      @OA\Property(
     *                          property="title",
     *                          type="string"
     *                      ),
     *                      @OA\Property(
     *                          property="letter",
     *                          type="string"
     *                      )
     *                  ),
     *                  @OA\Property(
     *                      property="city",
     *                      type="object",
     *                      @OA\Property(
     *                          property="id",
     *                          type="integer"
     *                      ),
     *                      @OA\Property(
     *                          property="title",
     *                          type="string"
     *                      ),
     *                      @OA\Property(
     *                          property="letter",
     *                          type="string"
     *                      )
     *                  ),
     *                  @OA\Property(
     *                      property="description",
     *                      type="string"
     *                  )
     *              ),
     *               @OA\Property(
     *                   property="candidates",
     *                   type="array",
     *                   @OA\Items(
     *                         @OA\Property(
     *                             property="id",
     *                             type="integer"
     *                         ),
     *                         @OA\Property(
     *                             property="name",
     *                             type="string"
     *                         ),
     *                         @OA\Property(
     *                             property="email",
     *                             type="string"
     *                         ),
     *                         @OA\Property(
     *                             property="phone",
     *                             type="string"
     *                         ),
     *                         @OA\Property(
     *                             property="image",
     *                             type="string"
     *                         ),
     *                         @OA\Property(
     *                             property="created_at",
     *                             type="string",
     *                             format="date-time"
     *                         ),
     *                         @OA\Property(
     *                             property="vewed_at",
     *                             type="string",
     *                             format="date"
     *                         )
     *                   )
     *               ),
     *               @OA\Property(
     *                   property="type",
     *                   type="object",
     *                   @OA\Property(
     *                       property="id",
     *                       type="integer"
     *                   ),
     *                   @OA\Property(
     *                       property="title",
     *                       type="string"
     *                   )
     *               ),
     *               @OA\Property(
     *                   property="occupation_area",
     *                   type="object",
     *                   @OA\Property(
     *                       property="id",
     *                       type="integer"
     *                   ),
     *                   @OA\Property(
     *                       property="title",
     *                       type="string"
     *                   )
     *               )
     *         )
     *      )
     * )
     */
    public function show(OpportunityService $opportunityService, $opportunity_id)
    {
        return $opportunityService->show($opportunity_id);
    }

    public function renew(OpportunityService $opportunityService, $opportunity_id)
    {
        return $opportunityService->renew($opportunity_id);
    }
}